requirements.txt
# Core Python Dependencies
setuptools>=80.0.0

# Django Core
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.3.1

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# Database
# psycopg2-binary==2.9.9  # Uncomment when PostgreSQL is installed

# Environment Variables
python-decouple==3.8

# Development Dependencies
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Production Dependencies
gunicorn==21.2.0
whitenoise==6.6.0

# Security
django-ratelimit==4.1.0
django-ipware==6.0.0
cryptography==41.0.7

# HTTP Requests (for testing)
requests==2.32.5

# Logging
structlog==23.2.0
